2017/10/04 Edwin Hastamorir Se hacen varias correcciones a la forma de generación del reporte, enviando un BigDecimal como parámetro en lugar de un BigInteger. Se reconfigura un botón para que sea posible volver a generar el reporte luego de cerrar la convocatoria.
Cosas pendientes:
En la aplicación móvil no se devuelve al listado de pruebas luego de hacer una 
evaluación.
El campo OBSERVACIÓN, al finalizar la evaluación, debe ser llenado de 
manera obligatoria, sin embargo, debe ser algo paramétrico que sea 
configurable por empresa.
No se puede visualizar el reporte luego de generado y antes de ser 
enviado por correo.
Se requiere otro reporte con las gráficas que resuman el resultado de 
la evaluación.
La aplicación móvil permite modificar la evaluación luego de haberla 
cerrado, lo cual es incorrecto. Es bueno que le permita verla pero no 
debe poder modificarla.
Cuando la convocatoria esta en estado "ALCANCE" el valor 
de la etiqueta "Empleados sin evaluar" puede quedar negativo. Hay que 
determinar si el problema es de la base de datos o de la aplicación
web.

Felicidades a Felipe por otro formidable proyecto.

+++ designereval-ejb/src/java/co/com/designer/eval/administrar/interfaz/IAdministrarInicio.java
@@ -32,7 +32,7 @@ public interface IAdministrarInicio {
 
     public List<Pruebas> obtenerPruebasEvaluado(String usuario, BigInteger secEmplConvo);
 
-    public boolean cerrarConvocatoria(BigInteger secConvocatoria);
+    public boolean cerrarConvocatoria(BigDecimal secConvocatoria);
 
     public List<Convocatorias> obtenerConvocatoriasAlcance(String usuario);
